backend/requirements-py313.txt
# Python 3.13 Compatible Requirements for Cement Plant Digital Twin Backend
# Updated versions that support Python 3.13

# Core FastAPI and server
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1

# Google AI and Cloud (updated versions)
google-generativeai==0.4.0
google-cloud-aiplatform==1.45.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Data processing (Python 3.13 compatible)
pandas==2.2.1
numpy==1.26.4
openpyxl==3.1.2

# API and validation
pydantic==2.6.1
httpx==0.27.0
aiofiles==23.2.1

# LangChain and AI (updated versions)
langchain==0.1.10
langchain-google-genai==0.0.11
langchain-community==0.0.25

# Vector database and embeddings
chromadb==0.4.24
sentence-transformers==2.5.1
faiss-cpu==1.8.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Additional utilities for Python 3.13
typing-extensions==4.9.0
pydantic-core==2.16.2
